chore(base): Move tests in the correct module.
This patch moves tests about `RoomNotableTags` in the newly created
`tags` module.
